Component renamed/Lint issues fixed according to new definitions
