--- src/components/FooterBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './footerBar';
