+++ src/components/badge/badge.stories.tsx
@@ -1,24 +1,18 @@
-import React from 'react';
-import BadgeUi from './index';
-import { storiesOf } from '@storybook/react';
+import { boolean, object, text, withKnobs } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
-import { withKnobs, text, boolean, object} from '@storybook/addon-knobs';
+import { storiesOf } from '@storybook/react';
+import BadgeUi from './index';
+import React from 'react';
 
 const stories = storiesOf('Badge', module);
-const BadgeDefaultStyle = {'color':'white','background-color':'green'};
+const BadgeDefaultStyle = { 'color':'white','background-color':'green' };
 
 
-stories.add('Badge with default value', () => {
-    return <BadgeUi badgeContent={'1'} color="primary"/>;
+stories.add('Badge with default value', () => { return <BadgeUi badgeContent={'1'} color="primary"/>;
 });
 
-stories.add('Badge with style', () => {
-    // const styleProps = {'background-color': backgroundColor};
-    const style = object('Style of Badge',BadgeDefaultStyle);
-    const content = text('Text', '007');
-    return <BadgeUi badgeContent={content} style={style}/>;
+stories.add('Badge with style', () => { const style = object('Style of Badge',BadgeDefaultStyle);const content = text('Text', '007');return <BadgeUi badgeContent={content} style={style}/>;
 });
 
-stories.add('Badge without knobs', () => {
-    return <BadgeUi badgeContent={'3'} style={BadgeDefaultStyle}/>;
+stories.add('Badge without knobs', () => {  return <BadgeUi badgeContent={'3'} style={BadgeDefaultStyle}/>;
 });
