+++ src/components/map/Map.component.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
 import { Gmaps, Marker } from 'react-gmaps';
 import  { IMap } from 'bluerain-ui-interfaces/Components/Map';
-
+import React from 'react';
 
 /**
  * The Map Component.
@@ -14,15 +13,15 @@ import  { IMap } from 'bluerain-ui-interfaces/Components/Map';
  */
 
 export interface IGoogleMaps {
-    width?: string,
-    height?: string,
-    zoom?: number,
-    draggable?: boolean,
-    params?: object,
+	width?: string,
+	height?: string,
+	zoom?: number,
+	draggable?: boolean,
+	params?: object,
 }
 
 
-const MapComponent:React.StatelessComponent<IMap & IGoogleMaps> = (props) =>(
+const MapComponent:React.StatelessComponent<IMap & IGoogleMaps> = (props) => (
     <Gmaps
         lat={props.region ? props.region.latitude : null}
         lng={props.region ? props.region.longitude : null}
@@ -38,12 +37,12 @@ const MapComponent:React.StatelessComponent<IMap & IGoogleMaps> = (props) =>(
 );
 
 MapComponent.defaultProps = {
-    params: {
-        v: '3.exp',
-        key: 'AIzaSyCgVYH02v3XYmCSIokoLOZzvUqJG6ZamMY'
-    },
-    zoom:5,
-    height:'600px'
+	params: {
+		v: '3.exp',
+		key: 'AIzaSyCgVYH02v3XYmCSIokoLOZzvUqJG6ZamMY'
+	},
+	zoom:5,
+	height:'600px'
 };
 
 export default MapComponent;
