+++ src/components/map/test/Map.stories.tsx
@@ -1,24 +1,25 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import {  text, number, boolean } from '@storybook/addon-knobs';
+
 import BR, { BlueRainProvider } from '@blueeast/bluerain-os';
+import {  boolean, number, text } from '@storybook/addon-knobs';
 import { IntlProvider } from 'react-intl';
+import { storiesOf } from '@storybook/react';
 import MapComponent from '../index';
+import React from 'react';
 const stories = storiesOf('Map', module);
 BR.boot({ renderApp:false });
 
 const onDragEnd = (e) => {
-    const { latLng } = e;
-    const latitude = latLng.lat();
-    const longitude = latLng.lng();
-    alert('Latitude and longitude');
+	const { latLng } = e;
+	const latitude = latLng.lat();
+	const longitude = latLng.lng();
+	alert('Latitude and longitude');
 };
 stories.add('Map with OnMarkDragEnd(withKnobs)', () => {
-    const draggable = boolean('draggable', true);
-    const zoom = number('zoom', 12);
-    const height = text('height', '500px');
-    const width = text('width', '100%');
-    return (
+	const draggable = boolean('draggable', true);
+	const zoom = number('zoom', 12);
+	const height = text('height', '500px');
+	const width = text('width', '100%');
+	return (
         <BlueRainProvider>
             <IntlProvider>
                 <MapComponent
@@ -33,8 +34,8 @@ stories.add('Map with OnMarkDragEnd(withKnobs)', () => {
         </BlueRainProvider>);
 });
 stories.add('Map with draggable(Knobs)', () => {
-    const draggable = boolean('draggable', false);
-    return (
+	const draggable = boolean('draggable', false);
+	return (
         <BlueRainProvider>
             <IntlProvider>
                 <MapComponent
@@ -46,8 +47,8 @@ stories.add('Map with draggable(Knobs)', () => {
 });
 
 stories.add('Map with zooming(Knobs)', () => {
-    const zoom = number('zoom', 12);
-    return (
+	const zoom = number('zoom', 12);
+	return (
         <BlueRainProvider>
             <IntlProvider>
                 <MapComponent
